Add full sorting to open claims
