+++ app/controllers/nsm/claims_controller.rb
@@ -1,20 +1,21 @@
 module Nsm
   class ClaimsController < ApplicationController
+    before_action :set_default_table_sort_options, only: %i[your open closed]
+
     def your
-      claims = Claim.pending_and_assigned_to(current_user)
-      pagy, filtered_claims = pagy(claims)
+      pagy, filtered_claims = order_and_paginate(Claim.pending_and_assigned_to(current_user))
       your_claims = filtered_claims.map { |claim| BaseViewModel.build(:your_claims, claim) }
 
       render locals: { your_claims:, pagy: }
     end
 
     def open
-      @pagy, claims = pagy(Claim.pending_decision)
-      @claims = claims.map { |claim| BaseViewModel.build(:all_claims, claim) }
+      @pagy, claims = order_and_paginate(Claim.pending_decision)
+      @claims = claims.map { |claim| BaseViewModel.build(:open_claims, claim) }
     end
 
     def closed
-      @pagy, claims = pagy(Claim.decision_made)
+      @pagy, claims = order_and_paginate(Claim.decision_made)
       @claims = claims.map { |claim| BaseViewModel.build(:assessed_claims, claim) }
     end
 
@@ -31,6 +32,18 @@ module Nsm
       else
         redirect_to your_nsm_claims_path, flash: { notice: t('.no_pending_claims') }
       end
+    end
+
+    private
+
+    def order_and_paginate(query)
+      pagy(Sorter.call(query, @sort_by, @sort_direction))
+    end
+
+    def set_default_table_sort_options
+      default = 'date_updated'
+      @sort_by = params.fetch(:sort_by, default)
+      @sort_direction = params.fetch(:sort_direction, 'descending')
     end
   end
 end
