--- app/views/nsm/claims/_open_table.html.erb
@@ -1,33 +1,0 @@
-<%=
-  govuk_table(html_attributes: {'data-module': 'moj-sortable-table'}) do |table|
-    table.with_caption(html_attributes: { class: "govuk-visually-hidden" }, text: t(".table_heading"))
-
-    table.with_head do |head|
-      head.with_row do |row|
-        row.with_cell(text: t('.laa_reference'))
-        row.with_cell(text: t('.firm_name'))
-        row.with_cell(text: t('.defendant_name'))
-        row.with_cell(text: t('.date_received'))
-        row.with_cell(text: t('.caseworker_name'))
-        row.with_cell(text: t('.status'))
-      end
-    end
-
-    table.with_body do |body|
-      claims.each do |claim|
-        body.with_row do |row|
-          row.with_cell do
-            govuk_link_to(claim.laa_reference, nsm_claim_claim_details_path(claim.id))
-          end
-          row.with_cell(text: claim.firm_name)
-          row.with_cell(text: claim.main_defendant_name)
-          row.with_cell(text: claim.date_created)
-          row.with_cell(text: claim.case_worker_name)
-          row.with_cell do
-            submission_state_tag(claim.submission)
-          end
-        end
-      end
-    end
-  end
-%>
