+++ app/views/nsm/claims/_table.html.erb
@@ -1,0 +1,33 @@
+<table class="govuk-table" data-module="moj-sortable-table" aria-describedby="page-title">
+  <thead class="govuk-table__head">
+    <tr class="govuk-table__row">
+      <% columns.keys.each_with_index do |column_key, index| %>
+        <%= table_header(column_key,
+                        "nsm.claims.table.header",
+                        index,
+                        path,
+                        @sort_by,
+                        @sort_direction) %>
+      <% end %>
+    </tr>
+  </thead>
+
+  <tbody class="govuk-table__body app-task-list__items">
+    <% claims.each do |claim| %>
+      <tr class="govuk-table__row app-task-list__item">
+        <% columns.values.each_with_index do |value_attribute, index| %>
+          <td class="govuk-table__cell">
+            <% if index.zero? %>
+              <%= link_to claim.send(value_attribute),
+                          nsm_claim_claim_details_path(claim.id),
+                          data: { turbo: false } %>
+            <% else %>
+              <%= claim.send(value_attribute) %>
+            <% end %>
+          </td>
+        <% end %>
+      </tr>
+    <% end %>
+  </tbody>
+</table>
+<%= render 'shared/pagination', pagy: pagy, item: t("nsm.claims.table.table_info_item") %>
