+++ app/views/nsm/claims/open.html.erb
@@ -3,7 +3,16 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-full">
     <h1 class="govuk-heading-xl"><%= t('.heading') %></h1>
-    <%= render 'open_table', { claims: @claims} %>
-    <%= render "shared/pagination", { pagy: @pagy, item: t('.table_info_item') } %>
+    <%= render 'table',
+                path: open_nsm_claims_path,
+                 columns: {
+                    laa_reference: :laa_reference,
+                    firm_name: :firm_name,
+                    client_name: :main_defendant_name,
+                    caseworker: :case_worker_name,
+                    date_updated: :date_updated,
+                    status: :state_tag
+                 },
+                claims: @claims, pagy: @pagy %>
   </div>
 </div>
